- Switch from float do double for better accuracy - high-performance implementions on embedded devices will need to do it with integers anyway. - Use lib/swt-3.1.jar instead of the Eclipse variable. Works better with different Eclipse installations. - Trim down the getWindowVariance method in TimeSeries again, because there is now a JUnit test. - Add getRe() and getIm() methods to Complex for more efficient computation of the power spectra in Coherence. - Implement Coherence and test it from ParallelPortPWMReader. Works! Yeah.
git-svn-id: file:///srv/svn/openuat-filtered-without-thirdparty/trunk/src/java/openuat@306 01b9256e-9321-419f-88ee-03bd83f56c6c

+++ src/org/eu/mayrhofer/sensors/SamplesSink.java
@@ -22,7 +22,7 @@ public interface SamplesSink {
 	 * @param index The index of this sample. All samples are required
 	 *              to be equally spaced.
 	 */
-	public void addSample(float sample, int index);
+	public void addSample(double sample, int index);
 
 	/** Should be called when it is detected that an active segment
 	 * starts, i.e. when it is detected that the source has become
